pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Serhii Charykov <laammaar@gmail.com>"]
description = ""
name = "exauth4"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
flask = "^2.3.3"
flask-cors = "^4.0.0"
flask-httpauth = "^4.8.0"
gevent = "^23.9.0.post1"
gunicorn = "^21.2.0"
pyjwt = "^2.8.0"
python = "^3.11"
python-dotenv = "^1.0.0"
redis = "^5.0.0"

[tool.poetry.group.dev.dependencies]
aiohttp = "^3.8.5"
mypy = "^1.5.1"
numpy = "^1.25.2"
psycopg = {extras = ["binary"], version = "^3.1.10"}
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-dotenv = "^0.5.2"
requests = "^2.31.0"
types-flask = "^1.1.6"
types-flask-cors = "^4.0.0.1"
types-redis = "^4.6.0.5"
